# .agents/resources/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "ColumnRefEdge": {
        "description": "Edge that maps a named output column to a named input.",
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "source_node": {
            "title": "Source Node",
            "type": "string"
          },
          "source_output": {
            "title": "Source Output",
            "type": "string"
          },
          "target_input": {
            "title": "Target Input",
            "type": "string"
          },
          "target_node": {
            "title": "Target Node",
            "type": "string"
          },
          "type": {
            "const": "column_ref",
            "default": "column_ref",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "source_node",
          "target_node",
          "source_output",
          "target_input"
        ],
        "title": "ColumnRefEdge",
        "type": "object"
      },
...
      "DisconnectOperation": {
        "description": "Remove an edge by id.",
        "properties": {
          "edge_id": {
            "title": "Edge Id",
            "type": "string"
          },
          "type": {
            "const": "disconnect",
            "default": "disconnect",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "edge_id"
        ],
        "title": "DisconnectOperation",
        "type": "object"
      },
      "DraftGraphProposalCreateRequest": {
        "description": "Draft-scoped proposal creation request.",
        "properties": {
          "base_revision": {
            "title": "Base Revision",
            "type": "integer"
          },
          "operations": {
            "items": {
              "discriminator": {
                "mapping": {
                  "add_node": "#/components/schemas/AddNodeOperation",
                  "connect": "#/components/schemas/ConnectOperation",
                  "disconnect": "#/components/schemas/DisconnectOperation",
                  "replace_graph": "#/components/schemas/ReplaceGraphOperation",
                  "update_parameters": "#/components/schemas/UpdateParametersOperation"
                },
                "propertyName": "type"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/AddNodeOperation"
                },
                {
                  "$ref": "#/components/schemas/ConnectOperation"
                },
                {
                  "$ref": "#/components/schemas/DisconnectOperation"
                },
                {
                  "$ref": "#/components/schemas/UpdateParametersOperation"
                },
                {
                  "$ref": "#/components/schemas/ReplaceGraphOperation"
                }
              ]
            },
            "title": "Operations",
            "type": "array"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": [
          "base_revision",
          "operations"
        ],
        "title": "DraftGraphProposalCreateRequest",
        "type": "object"
      },
...
      "GraphProposal": {
        "description": "Stored graph proposal.",
        "properties": {
          "base_revision": {
            "title": "Base Revision",
            "type": "integer"
          },
          "draft_id": {
            "title": "Draft Id",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "operations": {
            "items": {
              "discriminator": {
                "mapping": {
                  "add_node": "#/components/schemas/AddNodeOperation",
                  "connect": "#/components/schemas/ConnectOperation",
                  "disconnect": "#/components/schemas/DisconnectOperation",
                  "replace_graph": "#/components/schemas/ReplaceGraphOperation",
                  "update_parameters": "#/components/schemas/UpdateParametersOperation"
                },
                "propertyName": "type"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/AddNodeOperation"
                },
                {
                  "$ref": "#/components/schemas/ConnectOperation"
                },
                {
                  "$ref": "#/components/schemas/DisconnectOperation"
                },
                {
                  "$ref": "#/components/schemas/UpdateParametersOperation"
                },
                {
                  "$ref": "#/components/schemas/ReplaceGraphOperation"
                }
              ]
            },
            "title": "Operations",
            "type": "array"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": [
          "id",
          "draft_id",
          "base_revision",
          "operations"
        ],
        "title": "GraphProposal",
        "type": "object"
      },
      "GraphProposalApplyResponse": {
        "description": "Response returned after a proposal is applied.",
        "properties": {
          "draft_id": {
            "title": "Draft Id",
            "type": "string"
          },
          "graph": {
            "$ref": "#/components/schemas/GraphState"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "validation": {
            "$ref": "#/components/schemas/ValidationResult"
          }
        },
        "required": [
          "draft_id",
          "revision",
          "graph",
          "validation"
        ],
        "title": "GraphProposalApplyResponse",
        "type": "object"
      },
      "GraphProposalCreateRequest": {
        "description": "Request to create a proposal tied to a draft revision.",
        "properties": {
          "base_revision": {
            "title": "Base Revision",
            "type": "integer"
          },
          "draft_id": {
            "title": "Draft Id",
            "type": "string"
          },
          "operations": {
            "items": {
              "discriminator": {
                "mapping": {
                  "add_node": "#/components/schemas/AddNodeOperation",
                  "connect": "#/components/schemas/ConnectOperation",
                  "disconnect": "#/components/schemas/DisconnectOperation",
                  "replace_graph": "#/components/schemas/ReplaceGraphOperation",
                  "update_parameters": "#/components/schemas/UpdateParametersOperation"
                },
                "propertyName": "type"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/AddNodeOperation"
                },
                {
                  "$ref": "#/components/schemas/ConnectOperation"
                },
                {
                  "$ref": "#/components/schemas/DisconnectOperation"
                },
                {
                  "$ref": "#/components/schemas/UpdateParametersOperation"
                },
                {
                  "$ref": "#/components/schemas/ReplaceGraphOperation"
                }
              ]
            },
            "title": "Operations",
            "type": "array"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": [
          "draft_id",
          "base_revision",
          "operations"
        ],
        "title": "GraphProposalCreateRequest",
        "type": "object"
      },
      "GraphState": {
        "description": "Complete graph state with nodes and edges.",
        "properties": {
          "edges": {
            "items": {
              "discriminator": {
                "mapping": {
                  "column_ref": "#/components/schemas/ColumnRefEdge",
                  "positional": "#/components/schemas/PositionalEdge"
                },
                "propertyName": "type"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/ColumnRefEdge"
                },
                {
                  "$ref": "#/components/schemas/PositionalEdge"
                }
              ]
            },
            "title": "Edges",
            "type": "array"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/NodeState"
            },
            "title": "Nodes",
            "type": "array"
          },
          "published_inputs": {
            "items": {
              "$ref": "#/components/schemas/PublishedInput"
            },
            "title": "Published Inputs",
            "type": "array"
          },
          "published_outputs": {
            "items": {
              "$ref": "#/components/schemas/PublishedOutput"
            },
            "title": "Published Outputs",
            "type": "array"
          }
        },
        "required": [
          "nodes",
          "edges"
        ],
        "title": "GraphState",
        "type": "object"
      },
...
      "GraphValidationRequest": {
        "description": "Request to validate a graph in an optional workflow context.",
        "properties": {
          "graph": {
            "$ref": "#/components/schemas/GraphState"
          },
          "workflow_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workflow Name"
          }
        },
        "required": [
          "graph"
        ],
        "title": "GraphValidationRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OpenHandsContext": {
        "additionalProperties": false,
        "description": "Effective OpenHands configuration exposed to frontend callers.",
        "properties": {
          "available": {
            "title": "Available",
            "type": "boolean"
          },
          "deployment_mode": {
            "title": "Deployment Mode",
            "type": "string"
          },
          "host": {
            "title": "Host",
            "type": "string"
          },
          "port": {
            "title": "Port",
            "type": "integer"
          },
          "process_acknowledged": {
            "title": "Process Acknowledged",
            "type": "boolean"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "runtime": {
            "title": "Runtime",
            "type": "string"
          },
          "unsafe_webapp_features_enabled": {
            "title": "Unsafe Webapp Features Enabled",
            "type": "boolean"
          },
          "url": {
            "title": "Url",
            "type": "string"
          },
          "workspace": {
            "title": "Workspace",
            "type": "string"
          }
        },
        "required": [
          "available",
          "deployment_mode",
          "unsafe_webapp_features_enabled",
          "runtime",
          "host",
          "port",
          "url",
          "workspace",
          "process_acknowledged"
        ],
        "title": "OpenHandsContext",
        "type": "object"
      },
      "OpenHandsStatus": {
        "additionalProperties": false,
        "description": "Runtime status for the owned OpenHands process.",
        "properties": {
          "available": {
            "title": "Available",
            "type": "boolean"
          },
          "pid": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pid"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "running": {
            "title": "Running",
            "type": "boolean"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "required": [
          "available",
          "running"
        ],
        "title": "OpenHandsStatus",
        "type": "object"
      },
...
      "PositionalEdge": {
        "description": "Edge that connects nodes by positional index.",
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "positional_index": {
            "title": "Positional Index",
            "type": "integer"
          },
          "source_node": {
            "title": "Source Node",
            "type": "string"
          },
          "target_node": {
            "title": "Target Node",
            "type": "string"
          },
          "type": {
            "const": "positional",
            "default": "positional",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "source_node",
          "target_node",
          "positional_index"
        ],
        "title": "PositionalEdge",
        "type": "object"
      },
...
      "ValidationResult": {
        "description": "Result of validating a graph.",
        "properties": {
          "errors": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/GraphValidationError"
            },
            "title": "Errors",
            "type": "array"
          },
          "node_statuses": {
            "additionalProperties": {
              "$ref": "#/components/schemas/NodeStatus"
            },
            "default": {},
            "title": "Node Statuses",
            "type": "object"
          },
          "valid": {
            "title": "Valid",
            "type": "boolean"
          }
        },
        "required": [
          "valid"
        ],
        "title": "ValidationResult",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "BioImageFlow Server",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/graph": {
      "put": {
        "operationId": "validate_graph_endpoint_api_v1_graph_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/GraphState"
                  },
                  {
                    "$ref": "#/components/schemas/GraphValidationRequest"
                  }
                ],
                "title": "Body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Validate Graph Endpoint",
        "tags": [
          "graph"
        ]
      }
    },
...
    "/api/v1/graph/proposals": {
      "post": {
        "operationId": "create_graph_proposal_api_v1_graph_proposals_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GraphProposalCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphProposal"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Graph Proposal",
        "tags": [
          "graph"
        ]
      }
    },
    "/api/v1/graph/proposals/{proposal_id}/apply": {
      "post": {
        "operationId": "apply_graph_proposal_api_v1_graph_proposals__proposal_id__apply_post",
        "parameters": [
          {
            "in": "path",
            "name": "proposal_id",
            "required": true,
            "schema": {
              "title": "Proposal Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphProposalApplyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Apply Graph Proposal",
        "tags": [
          "graph"
        ]
      }
    },
    "/api/v1/graph/proposals/{proposal_id}/reject": {
      "post": {
        "operationId": "reject_graph_proposal_api_v1_graph_proposals__proposal_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "proposal_id",
            "required": true,
            "schema": {
              "title": "Proposal Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphProposal"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reject Graph Proposal",
        "tags": [
          "graph"
        ]
      }
    },
...
    "/api/v1/openhands/context": {
      "get": {
        "operationId": "openhands_context_api_v1_openhands_context_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenHandsContext"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Openhands Context",
        "tags": [
          "openhands"
        ]
      },
      "post": {
        "operationId": "receive_openhands_context_api_v1_openhands_context_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Payload",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Receive Openhands Context Api V1 Openhands Context Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Receive Openhands Context",
        "tags": [
          "openhands"
        ]
      }
    },
...
    "/api/v1/openhands/status": {
      "get": {
        "operationId": "openhands_status_api_v1_openhands_status_get",
        "parameters": [
          {
            "in": "query",
            "name": "launch",
            "required": false,
            "schema": {
              "default": false,
              "title": "Launch",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenHandsStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Openhands Status",
        "tags": [
          "openhands"
        ]
      }
    },
...
    "/api/v1/workflow-drafts/{draft_id}/agent-proposals": {
      "post": {
        "operationId": "create_draft_graph_proposal_api_v1_workflow_drafts__draft_id__agent_proposals_post",
        "parameters": [
          {
            "in": "path",
            "name": "draft_id",
            "required": true,
            "schema": {
              "title": "Draft Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftGraphProposalCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphProposal"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Draft Graph Proposal",
        "tags": [
          "workflow-drafts"
        ]
      }
    },
    "/api/v1/workflow-drafts/{draft_id}/agent-proposals/{proposal_id}/apply": {
      "post": {
        "operationId": "apply_draft_graph_proposal_api_v1_workflow_drafts__draft_id__agent_proposals__proposal_id__apply_post",
        "parameters": [
          {
            "in": "path",
            "name": "draft_id",
            "required": true,
            "schema": {
              "title": "Draft Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "proposal_id",
            "required": true,
            "schema": {
              "title": "Proposal Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphProposalApplyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Apply Draft Graph Proposal",
        "tags": [
          "workflow-drafts"
        ]
      }
    },
    "/api/v1/workflow-drafts/{draft_id}/agent-proposals/{proposal_id}/reject": {
      "post": {
        "operationId": "reject_draft_graph_proposal_api_v1_workflow_drafts__draft_id__agent_proposals__proposal_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "draft_id",
            "required": true,
            "schema": {
              "title": "Draft Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "proposal_id",
            "required": true,
            "schema": {
              "title": "Proposal Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphProposal"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reject Draft Graph Proposal",
        "tags": [
          "workflow-drafts"
        ]
      }
    },
...
  }
}